[CI] [87817,87819,87818]  <wifi> <perf> Optimize the ADC interface calls on the AP side within the SMP branch

+++ ap/include/driver/adc.h
@@ -392,7 +392,7 @@ adc_mode_t bk_adc_get_mode(void);
  * @return
  *    - float: volt
  */
-float bk_adc_data_calculate(UINT16 adc_val, UINT8 adc_chan);
+UINT16 bk_adc_data_calculate(UINT16 adc_val, UINT8 adc_chan);
 /**
  * @brief     Init the ADC GPIO
  *
